initBranch
