+++ script/htmlExports/sectionOne.js
@@ -1,0 +1,4 @@
+export const sectionOneHtml = `
+<h1 class="sectionTitles" id="sectionOneTitle">Setting a new standard for promises.</h1>
+<p class="sectionOneSmallText">We believe in a revolution within the DeFi space, and for that 
+trustworthy tools are needed to engage a trustless world of finance.</p>`
