+++ script/htmlExports/sectionThree.js
@@ -1,0 +1,42 @@
+export const sectionThreeHtml =         
+`<h1 class="sectionTitles" id="sectionThreeTitle">About us</h1>
+<p class="sectionOneSmallText">
+Our goal is to build a Web3 dApp for people to make social agreements and build trust in an anonymous environment.
+<br><br><br>
+The LOVA dApp provides a two step solution that any Web3 beginner can quickly put to use. The simplicity of LOVA revolutionizes the way of people can work and interact in Defi and beyond.
+<br><br><br>
+We are excited about the future of DeFi and are committed to building the tools and infrastructure necessary to make it a reality.
+Join us on this journey and be a part of shaping the future of finance.</p>
+
+<section class="sectionBottomPage">
+    <h2 class="sectionBottomTitle">Our Team</h1>
+    <figure class="teamImg">
+        <img class="imgImg" src="./imgs/jackS.webp"></img>
+        <p class="imgCardName"><b>Jack Smith Insulander</b></p>
+        <p class="imgCardText">Lead Developer<br><b>
+        <a href="https://github.com/jacksmithinsulander" class="personalLinks">Github</a><br>
+        <a href="https://www.linkedin.com/in/jack-smith-insulander-903220251/" class="personalLinks">LinkedIn</a></b></p>
+    </figure>
+    <figure class="teamImg">
+        <img class="imgImg" src="./imgs/felipeS.webp"></img>
+        <p class="imgCardName"><b>Felipe Stiernhoff</b></p>
+        <p class="imgCardText">Developer<br><b>
+        <a href="https://github.com/felipestiernhoff" class="personalLinks">Github</a><br>
+        <a href="https://www.linkedin.com/in/felipe-stiernhoff-558515123/" class="personalLinks">LinkedIn</a></b></p>
+    </figure>
+    <figure class="teamImg">
+        <img class="imgImg" src="./imgs/maxC.webp"></img>
+        <p class="imgCardName"><b>Max Cid</b></p>
+        <p class="imgCardText">Developer<br><b>
+        <a href="https://github.com/MaxCid" class="personalLinks">Github</a><br>
+        <a href="https://www.linkedin.com/in/maxcid/" class="personalLinks">LinkedIn</a></b></p>
+    </figure>
+    <figure class="teamImg">
+        <p class="imgCardName"><b>Philip Andelic</b></p>
+        <p class="imgCardText">Developer<br><b>
+        <a href="https://github.com/Gitmazter" class="personalLinks">Github</a><br>
+        <a href="https://www.linkedin.com/in/philip-andelic-9ba890123/" class="personalLinks">LinkedIn</a></b></p>
+        <img class="imgImg" src="./imgs/philipA.webp"></img>
+    </figure>
+</section>
+`;
