+++ script/htmlExports/sectionTwo.js
@@ -1,0 +1,12 @@
+export const sectionTwoHtml = `
+<h1 class="sectionTitles" id="sectionTwoTitle">How to</h1>
+<p class="sectionOneSmallText" id="howToGuideTxt">
+1. Install and log in to Metamask. <br><br>
+2. Visit our website and connect your Metamask wallet to our LOVA dApp. <br><br>
+3. Click on the "Create" button. <br> <br>
+4. Enter a title for your promise and deposit the desired amount of ETH. <br><br>
+5. Share the unique ID generated for your promise with the other participant. <br><br>
+6. The other participant must accept the promise within 10 minutes, after which the option to join will expire. <br><br>
+7. Both participants must sign the promise to confirm their commitment. <br><br>
+8. Once both participants have signed the promise, the funds will be released to the initial depositor. <br><br>
+    </p>`
