--- script/modules/connect.js
@@ -1,12 +1,0 @@
-export const connect = async () => {
-    if (typeof window.ethereum !== "undefined") {
-        await window.ethereum.request({
-            method: "eth_requestAccounts",
-        });
-        signer = provider.getSigner();
-        smartPromiseContract.connect(signer);
-        listenToEvent();
-    } else {
-        console.log("No metamask");
-    }
-};
